renaming files and constants after review
