+++ src/shared/modules/userFavorites/client/index.js
