+++ src/shared/modules/userFavorites/client/methods/create-many-user-favorites.js
