+++ src/shared/modules/userFavorites/client/methods/create-user-favorite.js
