+++ src/shared/modules/userFavorites/client/methods/get-all-user-favorites.js
