+++ src/shared/modules/userFavorites/client/methods/remove-many-user-favorites.js
