+++ src/shared/modules/userFavorites/client/methods/remove-user-favorite.js
