+++ src/shared/modules/userFavorites/client/methods/update-many-user-favorites.js
