+++ src/shared/modules/userFavorites/client/methods/update-user-favorite.js
